+++ abcmm-domain/src/test/java/com/aevoreth/abcmm/domain/playback/PlaybackSessionTest.java
@@ -139,6 +139,11 @@ class PlaybackSessionTest {
         }
 
         @Override
+        public void panic() {
+            stop();
+        }
+
+        @Override
         public void seek(Duration position) {
         }